fix(#3060): split test_calc_backend.py (1058 lines) into 4 focused test files
Moved shared client fixture to conftest.py and reorganized 81 tests into:
- test_app.py (TestAppStartup)
- test_inline_calcs.py (TestPressureDrop, TestFlowRate, TestODESolver,
  TestThermalProfile, TestSyngasWater, TestAcidGasDewpoint,
  TestSyngasWaterSanitize, TestScrubberAsFloat, TestPressureDropEdgeCases)
- test_mocked_routes.py (TestWGSReactorMocked, TestFlareRouterMocked,
  TestScrubberRouteMocked, TestBaghouseRouteMocked)
- test_protocols.py (TestProtocols, TestODEContracts)

Deleted original test_calc_backend.py. No test logic was modified.
